pyproject.toml
[tool.poetry]
name = "diplom-diplomych"
version = "0.1.0"
description = ""
authors = ["a.vaits <a.vaits@skai.online>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.14.0"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
loguru = "^0.7.2"
asyncpg = "^0.30.0"
pydantic = "^2.10.6"
pydantic-settings = "^2.8.1"
redis = "^5.2.1"
httpx = "^0.28.1"
psycopg2-binary = "^2.9.10"
fastapi = {extras = ["all"], version = "^0.115.11"}
sqladmin = "^0.20.1"
openai = "^1.66.3"
google-generativeai = ">=0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
